backend/requirements.txt
# Core
fastapi==0.115.7
uvicorn[standard]==0.34.0
python-multipart==0.0.18

# Database
sqlalchemy==2.0.39
psycopg2-binary==2.9.10
alembic==1.18.3

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dateutil==2.9.0

# Validation
pydantic==2.12.5
pydantic-settings==2.12.0
email-validator==2.2.0

# Utils
python-dotenv==1.0.1
